refactor: move sorting violations from a11y error into a11y results
